refactor: move useIntersectionObserver to ui-kit, add docs
