+++ apps/main/src/dao/components/ComboBoxSelectGauge/ComboBoxListChunk.tsx
@@ -1,7 +1,7 @@
 import type { ComboBoxSelectGaugeProps } from '@/dao/components/ComboBoxSelectGauge/types'
 import { RefObject, useRef } from 'react'
 import styled from 'styled-components'
-import useIntersectionObserver from 'ui/src/hooks/useIntersectionObserver'
+import useIntersectionObserver from '@ui-kit/hooks/useIntersectionObserver'
 import SelectGaugeListItem from '@/dao/components/ComboBoxSelectGauge/ComboBoxListItem'
 import { GaugeFormattedData } from '@/dao/types/dao.types'
