+++ apps/main/src/dex/components/PageDashboard/components/TableRow.tsx
@@ -2,7 +2,7 @@ import type { DashboardTableRowProps } from '@/dex/components/PageDashboard/type
 import { useRef } from 'react'
 import styled from 'styled-components'
 import { SORT_ID } from '@/dex/components/PageDashboard/utils'
-import useIntersectionObserver from '@ui/hooks/useIntersectionObserver'
+import useIntersectionObserver from '@ui-kit/hooks/useIntersectionObserver'
 import { Tr, Td } from '@ui/Table'
 import PoolLabel from '@/dex/components/PoolLabel'
 import TableCellBalances from '@/dex/components/PageDashboard/components/TableCellBalances'
