+++ apps/main/src/dex/components/PageDashboard/components/TableRowMobile.tsx
@@ -2,7 +2,7 @@ import type { DashboardTableRowProps } from '@/dex/components/PageDashboard/type
 import { useRef, useState } from 'react'
 import styled from 'styled-components'
 import { SORT_ID } from '@/dex/components/PageDashboard/utils'
-import useIntersectionObserver from 'ui/src/hooks/useIntersectionObserver'
+import useIntersectionObserver from '@ui-kit/hooks/useIntersectionObserver'
 import { Td, Tr } from '@ui/Table'
 import Box from '@ui/Box'
 import Icon from '@ui/Icon'
