+++ apps/main/src/dex/components/PagePoolList/components/TableRow.tsx
@@ -3,7 +3,7 @@ import type { CampaignRewardsMapper } from 'ui/src/CampaignRewards/types'
 import { t } from '@ui-kit/lib/i18n'
 import { Fragment, HTMLAttributes, useEffect, useRef, useState } from 'react'
 import { COLUMN_KEYS } from '@/dex/components/PagePoolList/utils'
-import useIntersectionObserver from '@ui/hooks/useIntersectionObserver'
+import useIntersectionObserver from '@ui-kit/hooks/useIntersectionObserver'
 import { CellInPool, Td, Tr } from '@ui/Table'
 import Box from '@ui/Box'
 import CampaignRewardsRow from '@/dex/components/CampaignRewardsRow'
