+++ apps/main/src/lend/components/PageMarketList/components/TableRowViewContentTable/TableRow.tsx
@@ -5,7 +5,7 @@ import { FilterType } from '@/lend/components/PageMarketList/utils'
 import { TITLE } from '@/lend/constants'
 import { _showContent } from '@/lend/utils/helpers'
 import useStore from '@/lend/store/useStore'
-import useIntersectionObserver from '@ui/hooks/useIntersectionObserver'
+import useIntersectionObserver from '@ui-kit/hooks/useIntersectionObserver'
 import { Tr, Td, CellInPool } from '@ui/Table'
 import CellLoanUserState from '@/lend/components/SharedCellData/CellLoanUserState'
 import CellLoanUserHealth from '@/lend/components/SharedCellData/CellLoanUserHealth'
