+++ apps/main/src/lend/components/PageMarketList/components/TableRowViewContentTable/TableRowMobile.tsx
@@ -1,7 +1,7 @@
 import type { TableCellProps, TableRowProps } from '@/lend/components/PageMarketList/types'
 import { ReactNode, useEffect, useRef, useState } from 'react'
 import useStore from '@/lend/store/useStore'
-import useIntersectionObserver from '@ui/hooks/useIntersectionObserver'
+import useIntersectionObserver from '@ui-kit/hooks/useIntersectionObserver'
 import TableRowMobileContent from '@/lend/components/PageMarketList/components/TableRowViewContentTable/TableRowMobileContent'
 import TableRowMobileHead from '@/lend/components/PageMarketList/components/TableRowViewContentTable/TableRowMobileHead'
 import { Tr } from '@ui/Table'
