+++ apps/main/src/loan/components/PageMarketList/components/TableRow/TableRow.tsx
@@ -4,7 +4,7 @@ import { Fragment, ReactNode, useRef } from 'react'
 import styled from 'styled-components'
 import networks from '@/loan/networks'
 import { TITLE } from '@/loan/constants'
-import useIntersectionObserver from '@ui/hooks/useIntersectionObserver'
+import useIntersectionObserver from '@ui-kit/hooks/useIntersectionObserver'
 import { CellInPool, Td, Tr } from '@ui/Table'
 import TokenLabel from '@/loan/components/TokenLabel'
 import TableCellRate from '@/loan/components/SharedCells/TableCellRate'
