+++ apps/main/src/loan/components/PageMarketList/components/TableRow/TableRowMobile.tsx
@@ -4,7 +4,7 @@ import { Dispatch, Fragment, ReactNode, SetStateAction, useRef } from 'react'
 import styled from 'styled-components'
 import networks from '@/loan/networks'
 import { TITLE } from '@/loan/constants'
-import useIntersectionObserver from '@ui/hooks/useIntersectionObserver'
+import useIntersectionObserver from '@ui-kit/hooks/useIntersectionObserver'
 import { Tr, CellInPool } from '@ui/Table'
 import Box from '@ui/Box'
 import Button from '@ui/Button'
