+++ packages/curve-ui-kit/src/shared/ui/DataTable.tsx
@@ -6,7 +6,7 @@ import TableBody from '@mui/material/TableBody'
 import Table from '@mui/material/Table'
 import TableRow from '@mui/material/TableRow'
 import { SizesAndSpaces } from '@ui-kit/themes/design/1_sizes_spaces'
-import { useIntersectionObserver } from 'ui'
+import useIntersectionObserver from '@ui-kit/hooks/useIntersectionObserver'
 import { Cell, Column, flexRender, Header, Row, useReactTable } from '@tanstack/react-table'
 import { ArrowDownIcon } from '@ui-kit/shared/icons/ArrowDownIcon'
 import { TransitionFunction } from '@ui-kit/themes/design/0_primitives'
