+++ packages/ui/src/hooks/index.ts
@@ -1,2 +1,1 @@
-export { default as useIntersectionObserver } from './useIntersectionObserver'
 export { default as useSlideTabState } from './useSlideTabState'
